# Adding Trie for WordPiece for faster encoding (Bert).