--- a/tokenizers/Cargo.toml
+++ b/tokenizers/Cargo.toml
@@ -33,6 +33,10 @@ harness = false
 name = "bert_benchmark"
 harness = false
 
+[[bench]]
+name = "load_benchmark"
+harness = false
+
 [dependencies]
 lazy_static = "1.4"
 rand = "0.7"